# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreatorReplyInputTone = typeof CreatorReplyInputTone[keyof typeof CreatorReplyInputTone];


export const CreatorReplyInputTone = {
  playful: 'playful',
  warm: 'warm',
  confident: 'confident',
  soft: 'soft',
  spicy: 'spicy',
} as const;

export interface CreatorReplyInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  question: string;
  adultConfirmed: boolean;
  /** @maxLength 80 */
  clientName?: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  personaName: string;
  tone: CreatorReplyInputTone;
  /** @maxLength 1000 */
  personaNotes?: string;
}

export interface CreatorReplyDraft {
  reply: string;
}

export interface AuthSessionStatus {
  authenticated: boolean;
}

export interface AccessCodeInput {
  /**
     * @minLength 4
     * @maxLength 4
     * @pattern ^[0-9]{4}$
     */
  code: string;
}

export type WhatsAppStatusConnection = typeof WhatsAppStatusConnection[keyof typeof WhatsAppStatusConnection];


export const WhatsAppStatusConnection = {
  disconnected: 'disconnected',
  connecting: 'connecting',
  awaiting_qr: 'awaiting_qr',
  connected: 'connected',
  error: 'error',
} as const;

export type WhatsAppStatusTone = typeof WhatsAppStatusTone[keyof typeof WhatsAppStatusTone];


export const WhatsAppStatusTone = {
  playful: 'playful',
  warm: 'warm',
  confident: 'confident',
  soft: 'soft',
  spicy: 'spicy',
} as const;

export interface WhatsAppStatus {
  connection: WhatsAppStatusConnection;
  /** @nullable */
  phoneNumber: string | null;
  /** @nullable */
  qrDataUrl: string | null;
  /** @nullable */
  lastError: string | null;
  autoReplyEnabled: boolean;
  /** @maxLength 80 */
  creatorName: string;
  tone: WhatsAppStatusTone;
  /** @nullable */
  personaNotes: string | null;
  /** @minimum 0 */
  approvedContactCount: number;
  /** @nullable */
  lastReplyAt: string | null;
}

export type WhatsAppSettingsInputTone = typeof WhatsAppSettingsInputTone[keyof typeof WhatsAppSettingsInputTone];


export const WhatsAppSettingsInputTone = {
  playful: 'playful',
  warm: 'warm',
  confident: 'confident',
  soft: 'soft',
  spicy: 'spicy',
} as const;

export interface WhatsAppSettingsInput {
  autoReplyEnabled: boolean;
  /** @maxLength 80 */
  creatorName: string;
  tone: WhatsAppSettingsInputTone;
  /**
     * @maxLength 1000
     * @nullable
     */
  personaNotes: string | null;
}

export interface WhatsAppContactInput {
  /**
     * @minLength 7
     * @maxLength 20
     */
  phoneNumber: string;
  /**
     * @maxLength 80
     * @nullable
     */
  displayName?: string | null;
  adultConfirmed: boolean;
}

export interface WhatsAppContact {
  id: number;
  phoneNumber: string;
  /** @nullable */
  displayName: string | null;
  adultConfirmed: boolean;
  createdAt: string;
}

export type WhatsAppContactList = WhatsAppContact[];

export type WhatsAppInboxMessageStatus = typeof WhatsAppInboxMessageStatus[keyof typeof WhatsAppInboxMessageStatus];


export const WhatsAppInboxMessageStatus = {
  pending: 'pending',
  sending: 'sending',
  uncertain: 'uncertain',
} as const;

export interface WhatsAppInboxMessage {
  /** @minimum 1 */
  id: number;
  phoneNumber: string;
  /**
     * @maxLength 80
     * @nullable
     */
  displayName: string | null;
  /** @maxLength 4000 */
  messageText: string;
  /**
     * @maxLength 4000
     * @nullable
     */
  replyDraft: string | null;
  status: WhatsAppInboxMessageStatus;
  receivedAt: string;
}

export type WhatsAppInboxMessageList = WhatsAppInboxMessage[];

export interface WhatsAppInboxReplyInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  reply: string;
}