# lib/api-zod/src/generated/types/whatsAppInboxMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WhatsAppInboxMessageStatus } from './whatsAppInboxMessageStatus';

export interface WhatsAppInboxMessage {
  /** @minimum 1 */
  id: number;
  phoneNumber: string;
  /**
     * @maxLength 80
     * @nullable
     */
  displayName: string | null;
  /** @maxLength 4000 */
  messageText: string;
  /**
     * @maxLength 4000
     * @nullable
     */
  replyDraft: string | null;
  status: WhatsAppInboxMessageStatus;
  receivedAt: Date;
}